requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-telegram-bot==20.7
aiofiles==23.2.1
pydantic==2.5.0

# PDF processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow>=10.3.0

# OCR
pytesseract==0.3.10
opencv-python==4.8.1.78

# ML/AI
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
regex==2023.10.3

# Jupyter
jupyter==1.0.0
notebook==7.0.6
matplotlib==3.8.2
seaborn==0.13.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Excel export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

